Arch: IFC export, move into separate module
